Move from 1999 to 2026 ... DCW/VPF is max size 32-bit integer
Drop usage of 'long' and just use 'int' or 'unsigned int'.
Actually, I'm not sure why I used 'long' in these types,
but I lucked out because in 1999 Visual Studio and Solaris
UNIX, and HPUX, a 'long' was (stil) just a 32-bit int.  So
I got away with it.

But jump to 2026 where CPUs are now 64-bit and ANSI compilers
now 'define' the 'long' to be 64-bit, then that clobbers the
parsing of the DCW/VPF data.

So this latches in JUST 32-bit integers for GMS.

Changes to be committed:
	modified:   include/gmsFile/gmsSpatialIndex.h
	modified:   include/gmsFile/gmsThematicIndex.h
	modified:   source/gmsFile/gmsSpatialIndex.cpp
	modified:   source/gmsFile/gmsThematicIndex.cpp

+++ C_source/include/gmsFile/gmsSpatialIndex.h
@@ -47,8 +47,8 @@
 
    typedef struct _nodeRecType
       {
-       long int offsetToNode;
-       long int numItemsInNode;
+       unsigned int offsetToNode;
+       unsigned int numItemsInNode;
       } nodeRecType;
 
    typedef struct _boundingRectangleType
